Add some adjustments to proposal data
