+++ assets/js/load-cities.js
@@ -4,7 +4,6 @@ import 'tom-select/dist/css/tom-select.default.min.css';
 document.addEventListener('DOMContentLoaded', () => {
     const stateElement  = document.getElementById('state');
     const cityElement   = document.getElementById('city');
-    const errorElement  = document.getElementById('error-message');
 
     stateElement.classList.remove('form-select');
     cityElement .classList.remove('form-select');
@@ -28,17 +27,6 @@ document.addEventListener('DOMContentLoaded', () => {
             .then(res => res.ok ? res.json() : [])
             .catch(() => []);
 
-    const showError = msg => {
-        if (!errorElement) return;
-        errorElement.textContent = msg;
-        errorElement.classList.remove('d-none');
-    };
-    const hideError = () => {
-        if (!errorElement) return;
-        errorElement.textContent = '';
-        errorElement.classList.add('d-none');
-    };
-
     const clearCities = () => {
         citySelect.clearOptions();
         citySelect.clear(true);
@@ -56,25 +44,6 @@ document.addEventListener('DOMContentLoaded', () => {
             citySelect.addOption({ value: c.id, text: c.name });
         });
         citySelect.refreshOptions(false);
-    });
-
-    citySelect.on('change', async value => {
-        hideError();
-        if (!value) return;
-
-        const name = citySelect.options[value]?.text?.trim();
-        if (!name) return;
-
-        const { exists } = await fetchData(
-            `/organizations/check-duplicate?name=${encodeURIComponent(name)}&cityId=${encodeURIComponent(value)}`
-        );
-
-        if (exists) {
-            showError(
-                `Este município já foi credenciado. Entre em contato com ${window.SNP_EMAIL} para mais informações.`
-            );
-            citySelect.clear(true);
-        }
     });
 
     const initialState = stateSelect.getValue();
