+++ templates/regmel/register/_partials/organization-data.html.twig
@@ -1,5 +1,9 @@
 <h4 class="mt-4">{{ 'organization_data' | trans }}</h4>
 
+<div class="alert alert-info">
+    Apenas para municípios e/ou <em>Arranjos Populacionais</em> a partir de 50.000 habitantes
+</div>
+
 <div class="row">
     <div class="col-6 mb-3">
         <label for="state" class="form-label required">{{ 'state' | trans }}</label>
