+++ templates/regmel/register/city.html.twig
@@ -66,5 +66,6 @@
         window.SNP_EMAIL = "{{ env.aurora.email.address }}";
     </script>
     <script type="module" src="{{ asset('js/load-cities.js') }}"></script>
+    <script type="module" src="{{ asset('js/check-duplicate-city.js') }}"></script>
     <script type="module" src="{{ asset('js/regmel/register.js') }}"></script>
 {% endblock %}
