+++ var/translations/index.js
@@ -178,6 +178,7 @@ export const EVIDENCES = {"id":"evidences","translations":{"messages":{"regmel":
 export const ELDERLY_PERSON = {"id":"elderly_person","translations":{"messages":{"regmel":"Pessoa idosa","pt-br":"Pessoa idosa","en":"Elderly Person","es":"Persona mayor"}}};
 export const EXECUTION_PERIOD = {"id":"execution_period","translations":{"messages":{"regmel":"Per\u00edodo de execu\u00e7\u00e3o","pt-br":"Per\u00edodo de execu\u00e7\u00e3o","en":"Execution Period","es":"Per\u00edodo de ejecuci\u00f3n"}}};
 export const EXPLORE_REPOSITORY = {"id":"explore_repository","translations":{"messages":{"regmel":"Conhe\u00e7a o reposit\u00f3rio","pt-br":"Conhe\u00e7a o reposit\u00f3rio","en":"Explore Repository","es":"Explorar Repositorio"}}};
+export const EXPORT_CSV = {"id":"export_csv","translations":{"messages":{"regmel":"Exportar CSV"}}};
 export const ELEVATOR = {"id":"elevator","translations":{"messages":{"regmel":"Elevador","pt-br":"Elevador","en":"Elevator","es":"Ascensor"}}};
 export const FAQ = {"id":"faq","translations":{"messages":{"regmel":"FaQ","pt-br":"FaQ","en":"FaQ","es":"FaQ"}}};
 export const FILE = {"id":"file","translations":{"messages":{"regmel":"Anexar arquivo","pt-br":"Anexar arquivo","en":"Attach file","es":"Adjuntar archivo"}}};
@@ -314,6 +315,7 @@ export const NO_AREAS = {"id":"no_areas","translations":{"messages":{"regmel":"N
 export const NO_ACCESSIBILITY = {"id":"no_accessibility","translations":{"messages":{"regmel":"Nenhuma acessibilidade registrada","pt-br":"Nenhuma acessibilidade registrada","en":"No accessibility registered","es":"Sin accesibilidad"}}};
 export const NO_SEALS = {"id":"no_seals","translations":{"messages":{"regmel":"Nenhum selo registrado","pt-br":"Nenhum selo registrado","en":"No seal","es":"Sin sello"}}};
 export const NONE_AREA_SPECIFIED = {"id":"none_area_specified","translations":{"messages":{"regmel":"Nenhuma \u00e1rea especificada","pt-br":"Nenhuma \u00e1rea especificada","en":"None area specified","es":"Ninguna zona especificada"}}};
+export const NO_DATA_FOUND = {"id":"no_data_found","translations":{"messages":{"regmel":"Nenhum dado encontrado"}}};
 export const NOT_INFORMED = {"id":"not_informed","translations":{"messages":{"regmel":"N\u00e3o informado","pt-br":"N\u00e3o informado","en":"Not informed","es":"No informado"}}};
 export const NOT_SENT = {"id":"not_sent","translations":{"messages":{"regmel":"N\u00e3o enviadas","pt-br":"N\u00e3o enviadas","en":"Not sent","es":"No enviado"}}};
 export const NOTICE = {"id":"notice","translations":{"messages":{"regmel":"EDITAL","pt-br":"EDITAL","en":"PUBLIC NOTICE","es":"AVISO PUBLICO"}}};
